添加日志处理

+++ manager-center/src/main/java/com/ysf/dao/UserMapper.java
@@ -5,13 +5,13 @@ import com.ysf.po.User;
 
 @Mapper
 public interface UserMapper {
-    int deleteByPrimaryKey(Integer userId);
+    int deleteByPrimaryKey(Long userId);
 
     int insert(User record);
 
     int insertSelective(User record);
 
-    User selectByPrimaryKey(Integer userId);
+    User selectByPrimaryKey(Long userId);
 
     int updateByPrimaryKeySelective(User record);
